Descriptor Needs Test
Adds Descriptors one add a time to the learning Method and iterates over "i" training and testing splits to get an average rms. Descriptors are either added in order they are listed in in the data file, or can be added in order from a file specified by FILENAME that has one number on each line that corresponds to a Descriptor index. This option is available if "F" is set to true. It continues to add Descriptors until it reaches the specified amount "Limit" or all Descriptors have been added