Add Descriptor Section
Useful if you want to see how RMS evolves over entire set of Descriptors, but learning method is slow. Adds desciptors to Data set in groups of length (m) and iterates over training-testing sets "r" times. Descriptors are either added in the order they are present in the data set or in the order specified in a secondary file if "F" is set to true.